refactor(marketing): drop barrel shims and thin announcement modules
Remove pure re-export index files, the deprecated SiteAnnouncements
shim, unused banner CSS var, and fold one-liner helpers into their
callers so imports point at real modules.
