+++ components/admin/site-announcements/AnnouncementFormDialog.tsx
@@ -16,11 +16,6 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import type {
-  AnnouncementEditor,
-  AnnouncementFormState,
-} from "@/lib/admin/siteAnnouncements";
-import { MultiSelectCombobox } from "@/components/ui/multi-select-combobox";
 import {
   DELAY_OPTIONS,
   LOCALE_OPTIONS,
@@ -29,8 +24,60 @@ import {
   SELECT_TRIGGER_CLASS,
   SITE_ANNOUNCEMENT_COUNTRY_OPTIONS,
   announcementUsesOverlay,
-} from "@/lib/constants/siteAnnouncements";
-import { FormField, FormFieldLabel, FormSettingRow } from "./FormPrimitives";
+  type AnnouncementEditor,
+  type AnnouncementFormState,
+} from "@/lib/admin/siteAnnouncements";
+import { MultiSelectCombobox } from "@/components/ui/multi-select-combobox";
+import { Label } from "@/components/ui/label";
+import { Switch } from "@/components/ui/switch";
+import { cn } from "@/lib/utils";
+
+function FormField({
+  children,
+  className,
+}: {
+  children: React.ReactNode;
+  className?: string;
+}) {
+  return <div className={cn("flex min-w-0 flex-col gap-1.5", className)}>{children}</div>;
+}
+
+function FormFieldLabel({
+  children,
+  required,
+}: {
+  children: React.ReactNode;
+  required?: boolean;
+}) {
+  return (
+    <Label className="inline-flex gap-0 font-sans text-sm font-medium leading-5 text-slate-700">
+      {children}
+      {required ? <span className="ml-0.5 text-red-500">*</span> : null}
+    </Label>
+  );
+}
+
+function FormSettingRow({
+  title,
+  description,
+  checked,
+  onCheckedChange,
+}: {
+  title: string;
+  description: string;
+  checked: boolean;
+  onCheckedChange: (value: boolean) => void;
+}) {
+  return (
+    <div className="flex items-center justify-between gap-3 py-2.5">
+      <div className="min-w-0">
+        <p className="text-sm font-medium text-slate-800">{title}</p>
+        <p className="text-[11px] text-slate-500">{description}</p>
+      </div>
+      <Switch checked={checked} onCheckedChange={onCheckedChange} className="shrink-0" />
+    </div>
+  );
+}
 
 interface AnnouncementFormDialogProps {
   editor: AnnouncementEditor;
