--- components/admin/site-announcements/FormPrimitives.tsx
@@ -1,50 +1,0 @@
-import { Label } from "@/components/ui/label";
-import { Switch } from "@/components/ui/switch";
-import { cn } from "@/lib/utils";
-
-export function FormField({
-  children,
-  className,
-}: {
-  children: React.ReactNode;
-  className?: string;
-}) {
-  return <div className={cn("flex min-w-0 flex-col gap-1.5", className)}>{children}</div>;
-}
-
-export function FormFieldLabel({
-  children,
-  required,
-}: {
-  children: React.ReactNode;
-  required?: boolean;
-}) {
-  return (
-    <Label className="inline-flex gap-0 font-sans text-sm font-medium leading-5 text-slate-700">
-      {children}
-      {required ? <span className="ml-0.5 text-red-500">*</span> : null}
-    </Label>
-  );
-}
-
-export function FormSettingRow({
-  title,
-  description,
-  checked,
-  onCheckedChange,
-}: {
-  title: string;
-  description: string;
-  checked: boolean;
-  onCheckedChange: (value: boolean) => void;
-}) {
-  return (
-    <div className="flex items-center justify-between gap-3 py-2.5">
-      <div className="min-w-0">
-        <p className="text-sm font-medium text-slate-800">{title}</p>
-        <p className="text-[11px] text-slate-500">{description}</p>
-      </div>
-      <Switch checked={checked} onCheckedChange={onCheckedChange} className="shrink-0" />
-    </div>
-  );
-}
