--- components/marketing/SiteAnnouncements.tsx
@@ -1,12 +1,0 @@
-/** @deprecated Prefer `@/components/marketing/site-announcements` */
-export {
-  ANNOUNCE_BANNER_HEIGHT_VAR,
-  SiteAnnouncementBanner,
-  SiteAnnouncementOverlays,
-  SiteAnnouncementPreviewOverlays,
-  SiteAnnouncementsProvider,
-  SiteHeaderSpacer,
-  SiteHeaderStack,
-  useSiteAnnouncementPreview,
-} from "./site-announcements";
-export type { AnnouncementType, SiteAnnouncement } from "./site-announcements";
