+++ components/marketing/site-announcements/index.ts
@@ -1,5 +1,3 @@
-export type { AnnouncementType, SiteAnnouncement } from "@/lib/marketing/siteAnnouncements";
-export { ANNOUNCE_BANNER_HEIGHT_VAR } from "@/lib/marketing/siteAnnouncements";
 export { SiteAnnouncementsProvider } from "./Provider";
 export { useSiteAnnouncementPreview } from "./context";
 export {
