+++ lib/marketing/siteAnnouncements/constants.ts
@@ -1,4 +1,3 @@
-export const ANNOUNCE_BANNER_HEIGHT_VAR = "--announce-banner-h";
 export const ANNOUNCE_BANNER_HEIGHT_PX = "36px";
 
 export const DISMISS_STORAGE_PREFIX = "fixera-announce-dismiss:";
@@ -13,3 +12,8 @@ export const SKIP_PATH_PREFIXES = [
   "/signup",
   "/register",
 ] as const;
+
+export function shouldSkipAnnouncements(pathname: string | null): boolean {
+  if (!pathname) return false;
+  return SKIP_PATH_PREFIXES.some((prefix) => pathname.startsWith(prefix));
+}
