--- lib/marketing/siteAnnouncements/index.ts
@@ -1,14 +1,0 @@
-export type { AnnouncementType, SiteAnnouncement } from "./types";
-export {
-  ANNOUNCE_BANNER_HEIGHT_VAR,
-  ANNOUNCE_BANNER_HEIGHT_PX,
-  PREVIEW_DURATION_MS,
-} from "./constants";
-export { resolveAnnouncementHref } from "./href";
-export { shouldSkipAnnouncements } from "./paths";
-export { fetchPublicSiteAnnouncements } from "./api";
-export { trackPromoClick, trackPromoView } from "./analytics";
-export {
-  isAnnouncementDismissed,
-  dismissAnnouncement,
-} from "./dismissStorage";
